+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [1.1.10]
+
+### Added
+
+- new `FD_ENTRY_ID` constant for entry id.
+
+### Removed
+
+- `processCommonSubmitActionFormData` function.
+- `getIntegrationApiPublicOutput` function.
+
 ## [1.1.9]
 
 ### Changed
@@ -64,6 +75,7 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 - Initial production release.
 
+[1.1.10]: https://github.com/infinum/eightshift-forms-utils/compare/1.1.9...1.1.10
 [1.1.9]: https://github.com/infinum/eightshift-forms-utils/compare/1.1.8...1.1.9
 [1.1.8]: https://github.com/infinum/eightshift-forms-utils/compare/1.1.7...1.1.8
 [1.1.7]: https://github.com/infinum/eightshift-forms-utils/compare/1.1.6...1.1.7